backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Cache
redis==5.0.1

# Auth (БЕЗ bcrypt - используем встроенный хеш)
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
